refactor(integration): split monolithic integration_test into domain files
